Dynamics 365 for Finance and Operations samples
Dynamics 365 for Finance and Operations fixes and sample code for Addresses and Contact information. When customers are imported from legacy systems or from Dynamics 365 for Sales, the addresses and contact information is non-editable. There is a field called IsExternallyMaintained that is set to Yes and this field sets all the addresses and contact information to non-editable (Primary and non-Primary addresses and contacts). Through this code, the non-primary addresses and contacts can be editable and the user can enter and edit non-primary addresses and contacts in Dynamics 365 for Finance and Operations.