Dynamics 365 for Finance and Operations fixes
When customers are imported from external systems into Dynamics 365 for Finance and Operations, the addresses and contact information (primary and non-primary) is non-editable for the imported customers. The reason for this is there is a flag called IsExternallyMaintained that is set to Yes and on setting this to Yes, the addresses and contact information (especially non-primary ones) are non-editable. This is a bug.
 Through the X++ code, the user can edit and enter non-primary addresses for customers in Dynamics 365 for Finance and Operations.