Added the X++ code for products and customers
Added the X++ Code in Finance and Operations for integration products between Dynamics 365 for Finance and Operations and Dynamics 365 for Sales. Similarly, for integrating customers between Dynamics 365 for Sales and Dynamics 365 for Finance and Operations. Just ensure that the relevant fields are mapped in the CDS Data Integration projects for the integration to work.